+++ src/main/java/com/openbravo/pos/sales/JPanelTicket.java
@@ -2950,7 +2950,7 @@ public abstract class JPanelTicket extends JPanel implements JPanelView, BeanFac
             try {
                 TicketInfo ticket = dlSales.loadTicket(
                         Integer.parseInt((m_config.getProperty("lastticket.type"))), 
-                        Integer.parseInt((m_config.getProperty("lastticket.number"))));
+                        Long.parseLong((m_config.getProperty("lastticket.number"))));
                 if (ticket == null) {
                     JFrame frame = new JFrame();
                     JOptionPane.showMessageDialog(frame,
